--- src/main/java/org/cloudfoundry/autosleep/business/remote/ApplicationIdentity.java
@@ -1,12 +1,0 @@
-package org.cloudfoundry.autosleep.business.remote;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ApplicationIdentity {
-    private String guid;
-
-    private String name;
-}
