--- src/main/java/org/cloudfoundry/autosleep/business/remote/CloudFoundryException.java
@@ -1,8 +1,0 @@
-package org.cloudfoundry.autosleep.business.remote;
-
-public class CloudFoundryException extends Exception {
-
-    public CloudFoundryException(Throwable cause) {
-        super(cause);
-    }
-}
