+++ src/test/java/org/cloudfoundry/autosleep/ui/servicebroker/service/AutosleepServiceInstanceServiceTest.java
@@ -8,7 +8,7 @@ import org.cloudfoundry.autosleep.dao.model.SpaceEnrollerConfig;
 import org.cloudfoundry.autosleep.dao.repositories.ApplicationRepository;
 import org.cloudfoundry.autosleep.dao.repositories.ServiceRepository;
 import org.cloudfoundry.autosleep.util.ApplicationLocker;
-import org.cloudfoundry.autosleep.business.GlobalWatcher;
+import org.cloudfoundry.autosleep.worker.GlobalWatcher;
 import org.cloudfoundry.autosleep.ui.servicebroker.service.parameters.ParameterReader;
 import org.cloudfoundry.autosleep.ui.servicebroker.service.parameters.ParameterReaderFactory;
 import org.cloudfoundry.autosleep.util.BeanGenerator;
